pyproject.toml
[project]
name = "auspol-kg"
version = "0.1.0"
description = "Knowledge graph extraction from Australian political media releases"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "beautifulsoup4>=4.12",
    "spacy>=3.7",
    "en-core-web-sm",
    "anthropic>=0.52",
    "networkx>=3.2",
    "pyvis>=0.3",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "cohere>=5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = ["pytest>=8.0"]

[tool.uv.sources]
en-core-web-sm = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl" }
